Made renaming s relocation work and added EpamCookiesConsentComponent
