+++ Selenium_pract_task/Helpers/EnvironmentVariableHelper.cs
