+++ Selenium_pract_task/Helpers/FileHelper.cs
