+++ Selenium_pract_task/Helpers/ScreenshotHelper.cs
