+++ Selenium_pract_task/POM/EpamAboutPage/EpamAboutPage.cs
@@ -1,11 +1,11 @@
 ﻿using OpenQA.Selenium;
 using Selenium_pract_task.Entities.AbstractEntities;
 
-namespace Selenium_pract_task.POM.EpamMainPage.EpamAboutPage
+namespace Selenium_pract_task.POM.EpamAboutPage
 {
     public class EpamAboutPage : AbstractPage
     {
-        public EpamAboutPage(IWebDriver driver) : base (driver)
+        public EpamAboutPage(IWebDriver driver) : base(driver)
         {
         }
