+++ Selenium_pract_task/POM/EpamInsightsPage/EpamResearchResultsPage/EpamResearchResultsPage.cs
@@ -1,7 +1,7 @@
 ﻿using OpenQA.Selenium;
 using Selenium_pract_task.Entities.AbstractEntities;
 
-namespace Selenium_pract_task.POM.EpamMainPage.EpamInsightsPage.EpamResearchResultsPage
+namespace Selenium_pract_task.POM.EpamInsightsPage.EpamResearchResultsPage
 {
     public class EpamResearchResultsPage : AbstractPage
     {
