+++ Selenium_pract_task/Tets_Cases/Validate_Title_Of_The_Article_Matches_With_Title_In_The_Carousel.cs
@@ -13,7 +13,7 @@ namespace Selenium_pract_task.Tets_Cases
         public void ValidateArticleMaches()
         {
             var epamMainPageContext = new EpamMainPageContext(driver);
-            epamMainPageContext.VerifyCookiesHandler()
+            epamMainPageContext.VerifyCookiesHandler<EpamMainPageContext>()
                 .ClickOnLinkInsightsLink()
                 .ClickButtonNextSlideContinuum(2)
                 .GetTextActiveSliderContinuum(out activeSliderText)
